+++ packages/renderer/src/features/settings/SettingsPage.tsx
@@ -3,6 +3,7 @@ import { useAtom } from 'jotai';
 import { useIpc } from '~/hooks';
 import { Icon } from '@mdi/react';
 import { lazy, useEffect} from 'react';
+import { APITab } from './components/APITab';
 import { AppTab } from './components/AppTab';
 import { AboutTab } from './components/AboutTab';
 import { YoutubeTab } from './components/YoutubeTab';
@@ -11,7 +12,7 @@ import { AdvancedTab } from './components/AdvancedTab';
 import { workingAtom, updatingAtom } from '~/atoms/app';
 import { DownloadsTab } from './components/DownloadsTab';
 import { Root, List, Trigger, Content } from '@radix-ui/react-tabs';
-import { mdiCogs, mdiTools, mdiYoutube, mdiDownload, mdiInformation, mdiApplicationCog } from '@mdi/js';
+import { mdiCogs, mdiTools, mdiYoutube, mdiDownload, mdiCodeBraces, mdiInformation, mdiApplicationCog } from '@mdi/js';
 import type { FC } from 'react';
 import type { TabsTriggerProps } from '@radix-ui/react-tabs';
 
@@ -70,25 +71,29 @@ export const SettingsPage = () => {
 					<TabButton value="app" title="Application" icon={mdiApplicationCog}/>
 					<TabButton value="downloads" title="Downloads" icon={mdiDownload}/>
 					<TabButton value="youtube" title="YouTube" icon={mdiYoutube}/>
+					<TabButton value="api" title="API" icon={mdiCodeBraces}/>
 					<TabButton value="advanced" title="Advanced" icon={mdiCogs}/>
 					<TabButton value="about" title="About" icon={mdiInformation}/>
 					{import.meta.env.DEV && <TabButton value="debug" title="Debug" icon={mdiTools}/>}
 				</List>
 				<div className="p-3 w-full overflow-y-auto [scrollbar-width:thin]">
 					<Content value="app">
-						<AppTab />
+						<AppTab/>
 					</Content>
 					<Content value="downloads">
-						<DownloadsTab />
+						<DownloadsTab/>
 					</Content>
 					<Content value="youtube">
-						<YoutubeTab />
+						<YoutubeTab/>
+					</Content>
+					<Content value="api">
+						<APITab/>
 					</Content>
 					<Content value="advanced">
-						<AdvancedTab />
+						<AdvancedTab/>
 					</Content>
 					<Content value="about">
-						<AboutTab />
+						<AboutTab/>
 					</Content>
 					{import.meta.env.DEV && (
 						<Content value="debug">